requirements.txt
altair==5.5.0
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
coverage==7.6.1
cycler==0.12.1
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.6.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.16.1
fonttools==4.55.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
identify==2.6.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.19.1
nodeenv==1.9.1
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pre-commit==3.8.0
protobuf==5.29.2
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyupgrade==3.17.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.6.8
seaborn==0.13.2
selenium==4.26.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
streamlit==1.41.1
tenacity==9.0.0
tokenize-rt==6.0.0
toml==0.10.2
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
